{{cookiecutter.repo_name}}/{{cookiecutter.package_name}}/apps/contact/forms.py
from captcha.fields import ReCaptchaField
from django import forms

from .models import ContactSubmission


class ContactForm(forms.ModelForm):

    captcha = ReCaptchaField()

    class Meta:
        model = ContactSubmission
        fields = '__all__'
