{{cookiecutter.repo_name}}/requirements.txt
argon2-cffi
beautifulsoup4==4.6.0
celery[redis]
CommonMark==0.7.4
cryptography
django>=1.11,<1.12
django-admin-sortable2
django-cachalot
django-compressor
django-extensions
django-flexible-images
django-historylinks
django-jet
django-jinja
django-lazy-image==0.4.0
django-recaptcha==1.4.0
{% if cookiecutter.redirects == 'yes' %}onespacemedia-cms-redirects{% endif %}
django-reversion==2.0.9
django-watson==1.3.1
django-webpack-loader
{% if cookiecutter.emails == 'yes' %}html2text==2018.1.9{% endif %}
Jinja2
lxml
onespacemedia-cms
onespacemedia-server-management
paramiko==1.16.0
Pillow
pygithub
python-memcached
rollbar
social-auth-app-django
sorl-thumbnail
requests
Werkzeug
{% if cookiecutter.geoip == 'yes' %}geoip{% endif %}
